♻️ Refactor metadata handling for static pages (home, about, faq)
