+++ app/faq/page.metadata.ts
@@ -1,0 +1,28 @@
+
+export const metadata = {
+    title: "FAQ | The Million Pixel Wall",
+    description: "Get answers to common questions about purchasing pixels, uploading images, and more.",
+    keywords: ["faq", "pixel questions", "help", "support", "million pixel wall"],
+    openGraph: {
+        title: "FAQ | The Million Pixel Wall",
+        description: "Everything you need to know about buying and managing your pixels.",
+        url: "https://the-million-pixel-wall.com/faq",
+        siteName: "The Million Pixel Wall",
+        images: [
+            {
+                url: "https://the-million-pixel-wall.com/og-faq.jpg",
+                width: 600,
+                height: 600,
+                alt: "FAQ – The Million Pixel Wall",
+            },
+        ],
+        locale: "en_US",
+        type: "website",
+    },
+    twitter: {
+        card: "summary_large_image",
+        title: "FAQ | The Million Pixel Wall",
+        description: "Clear answers to your pixel-related questions.",
+        images: ["https://the-million-pixel-wall.com/og-faq.jpg"],
+    },
+  };
