+++ app/not-found.tsx
@@ -1,0 +1,21 @@
+// app/not-found.tsx
+
+"use client";
+
+import Link from "next/link";
+import { Button } from "@/components/ui/button";
+
+export default function NotFoundPage() {
+  return (
+    <div className="flex flex-col items-center justify-center min-h-[60vh] text-center px-4">
+      <h1 className="text-5xl font-bold mb-4">404</h1>
+      <h2 className="text-2xl font-semibold mb-2">Page Not Found</h2>
+      <p className="text-muted-foreground mb-6">
+        Sorry, the page you’re looking for doesn’t exist.
+      </p>
+      <Button asChild>
+        <Link href="/">Go back home</Link>
+      </Button>
+    </div>
+  );
+}
